http get and get all request

+++ models/users.go
@@ -1,10 +1,10 @@
 package models
 
 type Users struct {
-    ID        int64    `json:"user_id"`
+    ID        int    `json:"user_id"`
     Username  string `json:"username"`
     Password  string `json:"password"`
-    Role      string `json:"role"` // e.g., "patient", "admin"
+    Role      string `json:"role"` // e.g., "user", "admin"
     FirstName string `json:"first_name"`
     LastName  string `json:"last_name"`
     Email     string `json:"email"`
